requirements.txt
functions-framework==3.*
pydantic==1.10.4
walrus==0.9.2
https://github.com/PADAS/er-client/releases/download/v1.1.0-alpha/earthranger_client-1.1.0-py3-none-any.whl
https://github.com/PADAS/smart-integrate-sdk/releases/download/v1.2.11.1/cdip_connector-1.2.11.1-py3-none-any.whl
gundi-client==0.2.2
opentelemetry-api==1.14.0
opentelemetry-sdk==1.14.0
opentelemetry-exporter-gcp-trace==1.3.0
opentelemetry-propagator-gcp
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-requests==0.35b0
opentelemetry-instrumentation-aiohttp-client==0.35b0
opentelemetry-instrumentation-confluent-kafka==0.35b0
opentelemetry-instrumentation-httpx==0.35b0
gcloud-aio-pubsub==5.2.0
pytest==7.2.1
pytest-asyncio==0.20.3
pytest-mock==3.10.0
